+++ x/ibctransfer/onrecv.go
@@ -47,7 +47,7 @@ func GetConvForIncomingCoins(
 		return nil, nil
 	}
 	decAmount := sdk.NewDecFromInt(sdk.NewIntFromUint64(data.Amount))
-	convAmountDec := IncreasePrecision(decAmount, po)
+	convAmountDec := IncreasePrecision(decAmount, int64(po))
 	convAmount := sdk.NewIntFromBigInt(convAmountDec.TruncateInt().BigInt())
 	// create converted and ibc tokens with corresponding denoms and amounts
 	convertToCoins := sdk.NewCoin(convertToDenomEntry.Denom, convAmount)
